Enabled test globbing. Minor style fixes.
All json test output is grouped into a single file.  Other applications
can sort it out.  Also added a second (mostly empty) test suite to test
if the json file is actually being written as a single file.  It is.  A
problem exists with this code in that the json can be written multiple
times. There is a refCount static variable in the AbstractZucchiniTest
to help account for this, but I don't know if it works yet.