Fixed a race condition with List instantiation.
Fixed a race condition in the AbstractZucchiniTest.runWith() method
wherein the List creation was not synchronized.  Creation is now
synchronized, along with List growth.  Additionally, the output from the
Zucchini test will now fail in a way that is visible to maven in the
case that the report failed to generate.