Added serialization option.  Code cleanup.
Added the option to force zucchini to use serialization by setting the
ZUCCHINI_SERIALIZATION environment variable to (y, yes, true, 1).
Removed some bad imports from the AbstractZuchiniTest class.  Changed
from a sysout for the ShutdownHook error log to using a logger and made
the input to the logger be getMessage() instead of a toString().