Fixes for Java7 and upgrade to Gson.
The previous Json package was compiled with Java8 and still managed to
work on some systems using Java7.  This should fix that problem of
inconsistency so that it works with all Java 7 systems.  Additionally,
the version of logback that was being used was old and caused problems
with OSX.  Lastly, the cucumber version was jumped to 1.2.4 in
preparation for future changes.

pom.xml - In addition to the previous errors that were fixed, the
spacing in the pom was fixed to be uniform.

AbstractZucchiniTest, TestContext, TestNGZucchiniRunner - Changed file
permissions, as these really shouldn't be executable.