Added documentation to recent changes.
Added documentation to all of the new classes and methods, as well as
some of the older methods that were still recently added, but lack
documentation.  At a later point, I will go through and add more
documentation to the classes that wrap Cucumber, because cucumber
provides no documentation for many of their public methods.