Race prevention with double-checked locking.
The addition of failed contexts had the potential to be unstable before.
This should help to make it more stable by ensuring that the actions
that are bundled with adding the failed context to the set are executed
only once.  I may need to break more rules at look at thread
pause/resume actions in the future.