Adds more protection for zucchini's execution.
Prevents threads from being killed when they are unable to catch the
ThreadDeath call.  This should help to prevent pre-kill or post-kill
(when the thread is running but is either past, or not yet in, a
scenario), though it doesn't help with potential blocking, as threads in
the pre state may not recognize that they have been killed, and will
then try to proceed.