Force barrier to be enabled for use.
Previously, all tests had the framework for the barrier sync
automatically enabled.  This change makes it so that the barrier sync
functionality needs to be enabled in order to be used, otherwise it will
skip barrier sync functions and throw if the barrier sync is called when
not supported.  The barrier sync is now disabled by default.