Shutdown hook to singleton.  Error log for setup.
The ZucchiniShutdownHook has be converted into a singleton instance,
which makes it easier to access it through the whole environment.  Along
with this is added logging to the ZucchiniShutdownHook that will
register whenever there is a problem with the setup and cleanup
procedures.  This should help to keep reporting more accurate.